Finished-HW7

+++ Homework/HW6/app/models.py
@@ -25,7 +25,7 @@ class Products(db.Model):
     SupplierID = db.Column(db.Integer, db.ForeignKey('suppliers.SupplierID'))
     CategoryID = db.Column(db.Integer, nullable=True)
     ProductName = db.Column(db.Integer, nullable=True)
-    UnitPrice = db.Column(db.Float(19,4), nullable=True)
+    UnitPrice = db.Column(db.Numeric, nullable=True)
     UnitsInStock = db.Column(db.Integer, nullable=True)
 
 # class Stats(db.Model):
